nrf52: clean scripts and restart node by powering the usb hub after programming to make sure the serial gets reconnected

+++ raspi/scripts/nrf52/serialdump.sh
@@ -2,8 +2,9 @@
 
 log_path=$1
 tty_path=`ls /dev/serial/by-id/usb-SEGGER_J-Link_*`
-killall -9 picocom
-killall -9 screen
+if pgrep picocom; then killall -9 picocom; fi; 
+if pgrep screen; then killall -9 screen; fi; 
+if pgrep contiki-timestamp; then killall -9 contiki-timestamp; fi;
 screen -wipe
 screen -dmS nrf52screen bash
 screen -S nrf52screen -X stuff "picocom -fh -b 115200 --imap lfcrlf $tty_path | ~/scripts/contiki-timestamp > $log_path\n"
